refactored root classes into relevant grouped namespaces
