+++ FlightJournal.Web/App_Start/IdentityConfig.cs
@@ -14,6 +14,7 @@ using System.Collections.Generic;
 using System.Data.Entity;
 using System.Threading.Tasks;
 using System.Web;
+using FlightJournal.Web.Configuration;
 using SendGrid;
 using Twilio;
