+++ FlightJournal.Web/App_Start/RouteConfig.cs
@@ -2,6 +2,8 @@
 using System.Security.Policy;
 using System.Web.Mvc;
 using System.Web.Routing;
+using FlightJournal.Web.Constraints;
+using FlightJournal.Web.Validators;
 
 namespace FlightJournal.Web{
     public class RouteConfig {
