+++ FlightJournal.Web/Extensions/HttpRequestBaseExtension.cs
@@ -1,28 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using FlightJournal.Web.Models;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin.Security;
-using System;
-using System.Collections.Generic;
-using System.Data.Entity;
-using System.Data.Entity.ModelConfiguration.Configuration;
-using System.Globalization;
-using FlightJournal.Web.Models;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin.Security;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Web;
-using System.Web.Mvc;
+﻿using System.Web;
 using FlightJournal.Web.Controllers;
+using FlightJournal.Web.Models;
 
-namespace FlightJournal.Web
+namespace FlightJournal.Web.Extensions
 {
     public static class HttpRequestBaseExtension
     {
