+++ FlightJournal.Web/Extensions/IPrincipalExtension.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Principal;
-using System.Text;
-using System.Threading.Tasks;
-using System.Web;
+﻿using System.Security.Principal;
 using FlightJournal.Web.Controllers;
 using FlightJournal.Web.Models;
 
-namespace FlightJournal.Web
+namespace FlightJournal.Web.Extensions
 {
     public static class IPrincipalExtension
     {
