+++ FlightJournal.Web/Validators/DatePathValidator.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Web;
 
-namespace FlightJournal.Web
+namespace FlightJournal.Web.Validators
 {
     public interface IDateRouteValidator
     {
