+++ FlightJournal.Web/Views/Club/ClubPresentation.cshtml
@@ -1,5 +1,6 @@
 ﻿@using FlightJournal.Web
 @using FlightJournal.Web.Controllers
+@using FlightJournal.Web.Extensions
 @model IEnumerable<FlightJournal.Web.Models.Club>
 @using (Html.BeginForm("SetCurrentClub", "Club", FormMethod.Post))
 {
