+++ FlightJournal.Web/Views/Flight/Details.cshtml
@@ -1,4 +1,5 @@
 ﻿@using FlightJournal.Web
+@using FlightJournal.Web.Extensions
 @model FlightJournal.Web.Models.Flight
 @{
     ViewBag.Title = "Details";
