+++ FlightJournal.Web/Views/Flight/Edit.cshtml
@@ -1,4 +1,5 @@
 ﻿@using FlightJournal.Web
+@using FlightJournal.Web.Extensions
 @model FlightJournal.Web.Models.Flight
 @{
     ViewBag.Title = "Edit";
