+++ FlightJournal.Web/Views/Flight/_changeHistory.cshtml
@@ -1,4 +1,5 @@
 ﻿@using FlightJournal.Web
+@using FlightJournal.Web.Extensions
 @model IEnumerable<FlightJournal.Web.Models.FlightVersionHistory>
 
 @helper TruncateDate(DateTime input)
