+++ FlightJournal.Web/Views/Home/Administration.cshtml
@@ -1,4 +1,5 @@
 @using FlightJournal.Web
+@using FlightJournal.Web.Extensions
 @{
     ViewBag.Title = "Administration";
 }
