+++ FlightJournal.Web/Views/Home/Index.cshtml
@@ -17,12 +17,16 @@
 </div>
 <div class="alert alert-warning">
     @Html.Action("ManagePilotBindingPartial", "Manage")
-    #TODO: Refactorere så Extensions får deres egen subfolder i projektet i stedet for roden
+    #TODO: Validerer link bygningen på Report Controller ved datoer - når der ikke er valgt klub går galt i byen med stien (kan ses på den global stat rapport ved valg af global klub)
     #TODO: Valider besked når der er manglende match på pilot og besked om hvordan man får et match<br/>
     f.eks. Herved får man adgang til startlisten og oprettelse og redigering af flyvninger foretaget på ens klubfly, startklub eller piloter i ens klub.<br/>
     #TODO: Validere at opret bruger processen har en god og pæn validering af bruger proces<br/>
     #TODO: Detaljeret visninger på flyvning mangler styling<br/>
     #TODO: Listen over felter på rediger flyvning mangler styling<br/>
+    #TODO: Validere proces for administrator med administration af managere<br />
+    #TODO: Validere proces for manager med administration af piloter<br/>
+    #TODO: Validere proces for manager med administration af editor<br />
+    #TODO: Validere proces for editor med administration af landingspladser og fly
 </div>
 
     <h2>Projekt fælles startliste</h2>
