+++ FlightJournal.Web/Views/Shared/_Layout.cshtml
@@ -1,5 +1,6 @@
 ﻿@using FlightJournal.Web
 @using FlightJournal.Web.Controllers
+@using FlightJournal.Web.Extensions
 @using FlightJournal.Web.Models
 <!DOCTYPE html>
 <html>
