+++ FlightJournal.Web/Views/Shared/_LoginPartial.cshtml
@@ -1,4 +1,5 @@
 @using FlightJournal.Web
+@using FlightJournal.Web.Extensions
 @using FlightJournal.Web.Models
 @using Microsoft.AspNet.Identity
 @if (Request.IsAuthenticated) {
