+++ FlightJournal.Web/Web.config
@@ -5,7 +5,7 @@
   -->
 <configuration>
   <configSections>
-    <section name="serviceCredentials" type="FlightJournal.Web.ServiceCredentialsConfigurationSection, FlightJournal.Web, Version=1.0.0.0, Culture=neutral" />
+    <section name="serviceCredentials" type="FlightJournal.Web.Configuration.ServiceCredentialsConfigurationSection, FlightJournal.Web, Version=1.0.0.0, Culture=neutral" />
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
     
   <!-- For more information on Entity Framework configuration, visit http://go.microsoft.com/fwlink/?LinkID=237468 --></configSections>
